Refactor LibraryAnalyser trait to use ApiRepresentation and update RustAnalyser implementation. Remove ApiDefinitions struct from types.rs and define it in rust/mod.rs with ApiRepresentation trait implementation. Adjust function signatures accordingly for improved type safety.
