+++ src/analysers.rs
@@ -1,22 +1,26 @@
 use crate::languages::rust::RustAnalyser;
 
 use crate::error::LaibraryError;
-use crate::types::{ApiDefinitions, PackageMetadata, SourceFile};
+use crate::types::{ApiRepresentation, PackageMetadata, SourceFile};
 use std::path::Path;
 
 pub trait LibraryAnalyser {
+    type Api: ApiRepresentation;
+
     fn extract_metadata(&self, path: &Path) -> Result<PackageMetadata, LaibraryError>;
     fn parse_source(&self, path: &Path) -> Result<Vec<SourceFile>, LaibraryError>;
-    fn extract_public_api(&self, sources: &[SourceFile]) -> Result<ApiDefinitions, LaibraryError>;
+    fn extract_public_api(&self, sources: &[SourceFile]) -> Result<Self::Api, LaibraryError>;
     fn generate_documentation(
         &self,
         metadata: &PackageMetadata,
-        api: &ApiDefinitions,
+        api: &Self::Api,
     ) -> Result<String, LaibraryError>;
 }
 
 /// Get an analyser for the specified language
-pub fn get_analyser(language: &str) -> Result<Box<dyn LibraryAnalyser>, LaibraryError> {
+pub fn get_analyser(
+    language: &str,
+) -> Result<Box<dyn LibraryAnalyser<Api = impl ApiRepresentation>>, LaibraryError> {
     match language {
         "rust" => Ok(Box::new(RustAnalyser::new())),
         _ => Err(LaibraryError::UnsupportedLanguage(language.to_string())),
