+++ src/languages/rust/impl_extractor.rs
@@ -1,5 +1,6 @@
+use super::ApiDefinitions;
 use crate::error::LaibraryError;
-use crate::types::{ApiDefinitions, SourceFile};
+use crate::types::SourceFile;
 use tree_sitter::{Node, Parser};
 use tree_sitter_rust::LANGUAGE;
